Add chapters scene
This removes the comics on the home page and replaces them with titles. Images will be needed here. Should we use the earliest issue we have?

The chapters scene does not sort comics yet, but should later on, tbd if we do this in the backend or front. Idealy the id row should match the issue number, would make it nice and clean.